+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add deactivate/active logic when using resources on shared storage.
+
 ## [0.13.0] - 2021-05-12
 
 ### Added